Refactor LLMJ project: added unified script, improved README, clean up repo, and included example outputs
