+++ scripts/LLMJ-5-eco-o3-mini/1_o3-mini_orkg_d1_b1/scores.csv
@@ -1,0 +1,2 @@
+report_stem,config,depth,breadth,rigor,innovation,gap,overall,depth_raw,breadth_raw,rigor_raw,innovation_raw,gap_raw,overall_raw,report_path
+1_o3-mini_orkg_d1_b1,d1_b1,0.49,0.62,0.62,0.59,0.62,0.588,49,62,62,59,62,59,/Users/mikeashleycedric/Documents/GitHub/deep-research/scripts/test-reports/ecology/1_o3-mini_orkg_d1_b1.md
