+++ scripts/LLMJ-5-eco-o3-mini/1_o3-mini_orkg_d1_b1_repo_compatible_row.csv
@@ -1,0 +1,2 @@
+config,depth_score,breadth_score,rigor_score,innovation_score,gap_score,overall_quality
+d1_b1,0.53,0.64,0.59,0.58,0.67,0.602
