+++ scripts/LLMJ-5-eco-o3-mini/1_o3-mini_orkg_d1_b1_scores_only.json
@@ -1,0 +1,8 @@
+{
+  "depth": 53,
+  "breadth": 64,
+  "rigor": 59,
+  "innovation": 58,
+  "gap": 67,
+  "overall_computed": 60
+}
