+++ scripts/LLMJ-5-eco-o3-mini/1_o3-mini_orkg_d1_b4/scores.csv
@@ -1,0 +1,2 @@
+report_stem,config,depth,breadth,rigor,innovation,gap,overall,depth_raw,breadth_raw,rigor_raw,innovation_raw,gap_raw,overall_raw,report_path
+1_o3-mini_orkg_d1_b4,d1_b4,0.53,0.72,0.51,0.73,0.67,0.632,53,72,51,73,67,63,/Users/mikeashleycedric/Documents/GitHub/deep-research/scripts/test-reports/ecology/1_o3-mini_orkg_d1_b4.md
