+++ scripts/LLMJ-5-eco-o3-mini/1_o3-mini_orkg_d4_b1/scores.csv
@@ -1,0 +1,2 @@
+report_stem,config,depth,breadth,rigor,innovation,gap,overall,depth_raw,breadth_raw,rigor_raw,innovation_raw,gap_raw,overall_raw,report_path
+1_o3-mini_orkg_d4_b1,d4_b1,0.66,0.91,0.58,0.69,0.69,0.706,66,91,58,69,69,71,/Users/mikeashleycedric/Documents/GitHub/deep-research/scripts/test-reports/ecology/1_o3-mini_orkg_d4_b1.md
