+++ scripts/LLMJ-5-eco-o3-mini/batch_config_summary.csv
@@ -1,0 +1,5 @@
+config,n_docs,depth_mean,depth_std,breadth_mean,breadth_std,rigor_mean,rigor_std,innovation_mean,innovation_std,gap_mean,gap_std,overall_mean,overall_std
+d1_b1,1,0.49,0.0,0.62,0.0,0.62,0.0,0.59,0.0,0.62,0.0,0.588,0.0
+d1_b4,1,0.53,0.0,0.72,0.0,0.51,0.0,0.73,0.0,0.67,0.0,0.632,0.0
+d4_b1,1,0.66,0.0,0.91,0.0,0.58,0.0,0.69,0.0,0.69,0.0,0.706,0.0
+d4_b4,0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0
