+++ scripts/LLMJ-5-eco-o3-mini/batch_reports.csv
@@ -1,0 +1,4 @@
+report_stem,config,depth,breadth,rigor,innovation,gap,overall,report_path
+1_o3-mini_orkg_d1_b1,d1_b1,0.49,0.62,0.62,0.59,0.62,0.588,/Users/mikeashleycedric/Documents/GitHub/deep-research/scripts/test-reports/ecology/1_o3-mini_orkg_d1_b1.md
+1_o3-mini_orkg_d1_b4,d1_b4,0.53,0.72,0.51,0.73,0.67,0.632,/Users/mikeashleycedric/Documents/GitHub/deep-research/scripts/test-reports/ecology/1_o3-mini_orkg_d1_b4.md
+1_o3-mini_orkg_d4_b1,d4_b1,0.66,0.91,0.58,0.69,0.69,0.706,/Users/mikeashleycedric/Documents/GitHub/deep-research/scripts/test-reports/ecology/1_o3-mini_orkg_d4_b1.md
